Fix token masking text selection
Tokens have a caret listener that will update the token to show text selection when the selection is completed but not during the selection itself, causing the token highlight to mask the region of text selection until the mouse is released. Add a mouse motion listener to tokens so that the token will update while dragging to select text, allowing the selection to be seen on top of the token itself.